+++ app.py
@@ -1,178 +1,17 @@
+corpus_source = "" # If empty, app will default to using the textbook PDF
+
 import os
-import time
+os.environ['CORPUS_SOURCE'] = corpus_source
+
 import subprocess
-import streamlit as st
-from roman import toRoman
-from streamlit_pdf_viewer import pdf_viewer
-from inference import chat_completion
-from statistics import (
-    init_db, 
-    init_user_session, 
-    update_user_session, 
-    insert_conversation, 
-    get_statistics,
-    toggle_correctness
-)
-
-init_db()  # Initialize the database
-
-def serve_pdf():
-    """Used to open PDF file when a citation is clicked"""
-    TEXTBOOK_PATH = os.path.join(os.path.dirname(__file__), "data", "textbook")
-    file = st.query_params.get("file")
-
-    # inference.py adjusts the pages for us
-    # Ex: Chapter 1 at pg 34 in the PDF will have a page value of 34 here
-    page = max(int(st.query_params.get("page", "1")), 1)
-    adjusted_page = page - 33 # Chapter 1 will be assigned page 1 but we have to handle pages before it
-    # For pages before Chapter 1
-    if adjusted_page < 1: # For pages before Chapter 1
-        # Display cover if first page else display roman numerals
-        adjusted_page = "Cover" if (page - 1) == 0 else toRoman(page - 1) # Display cover if first page else display roman numerals
-
-    if file:
-        pdf_path = os.path.join(TEXTBOOK_PATH, file)
-        if os.path.exists(pdf_path):
-            with st.spinner(f"Loading page {adjusted_page} of the textbook..."):
-                pdf_viewer(
-                    pdf_path,
-                    width=700,
-                    height=1000,
-                    pages_to_render=[page],
-                    scroll_to_page=page
-                )
-        else:
-            st.error(f"PDF file not found at {pdf_path}")
-    else:
-        st.error("No PDF file specified in query parameters")
-
-
-def update_and_display_statistics():
-    """Updates statistics report in the left sidebar"""
-    stats = get_statistics()
-    st.session_state.statistics = stats
-
-    statistics = [
-        f"Number of questions: {stats['num_questions']}",
-        f"Number of correct answers: {stats['num_correct']}",
-        f"Number of incorrect answers: {stats['num_incorrect']}",
-        f"User engagement metrics: {stats['user_engagement']:.2f} seconds",
-        f"Response time analysis: {stats['avg_response_time']:.2f} seconds",
-        f"Accuracy rate: {stats['accuracy_rate']:.2f}%",
-        f"Satisfaction rate: {stats['satisfaction_rate']:.2f}%",
-        f"Common topics or keywords",
-        f"Improvement over time",
-        f"Feedback summary",
-        f"Statistics per day and overall"
-    ]
-
-    st.sidebar.markdown("<h1 class='title-stat'>Statistics Reports</h1>", unsafe_allow_html=True)
-    for stat in statistics:
-        st.sidebar.markdown(f"""
-            <div class='btn-stat-container'>
-                <a href="#" class="btn-stat">{stat}</a>
-            </div>
-        """, unsafe_allow_html=True)
-
-def handle_feedback(conversation_id):
-    """Handle feedback button click"""
-    feedback_value = st.session_state.get(f"feedback_{conversation_id}", None)
-    if feedback_value is not None:
-        toggle_correctness(conversation_id, feedback_value == 1)
-        update_user_session(st.session_state.user_id)
-
-def main():
-    """Main Streamlit app logic."""
-    # Create the title
-    header = st.container()
-    header.title("Textbook Chatbot")
-    header.write("""<div class='fixed-header'/>""", unsafe_allow_html=True)
-
-    # Load PDF
-    if "view" in st.query_params and st.query_params["view"] == "pdf":
-        serve_pdf()
-    # Load Homepage
-    else:
-        # Load the CSS file
-        def load_css(file_name):
-            with open(file_name) as f:
-                st.markdown(f'<style>{f.read()}</style>', unsafe_allow_html=True)
-        load_css("Styles/style.css")
-
-        # Create user session
-        if "user_id" not in st.session_state:
-            st.session_state.user_id = init_user_session()
-            print(f"Creating user#{st.session_state.user_id}")
-        
-        # Create sidebar
-        st.sidebar.empty()
-        update_and_display_statistics()
-
-        # load user/assistant messages and feedback icons when appropriate
-        if "messages" not in st.session_state:
-            st.session_state.messages = []
-        for message in st.session_state.messages:
-            if message["role"] == "assistant":
-                st.markdown(f"<div class='assistant-message'>{message['content']}</div>", unsafe_allow_html=True)
-                conversation_id = message.get("conversation_id", None)
-                if conversation_id:
-                    st.caption("Was this response helpful?")
-                    feedback = st.feedback(
-                        "thumbs",
-                        key=f"feedback_{conversation_id}",
-                        on_change=handle_feedback,
-                        kwargs={"conversation_id": conversation_id}
-                    )
-            else:
-                st.markdown(f"<div class='user-message'>{message['content']}</div>", unsafe_allow_html=True)
-
-        # Handle user prompts
-        if prompt := st.chat_input("Ask your question?"):
-            st.markdown(f"<div class='user-message'>{prompt}</div>", unsafe_allow_html=True)
-
-            # Model inference with streaming
-            response_container = st.empty()  # Placeholder for the assistant's response
-            with st.spinner("Generating response..."):
-                start_time = time.time()
-                response = ""  # To accumulate the response in pieces
-                for partial_response, model_name in chat_completion(prompt):  # Streaming from the model
-                    response += partial_response  # Accumulate partial responses
-                    response_container.markdown(f"<div class='assistant-message'>{response}</div>", unsafe_allow_html=True)
-                end_time = time.time()
-                response_time = int((end_time - start_time)) # seconds
-
-            # Add conversation to DB
-            conversation_id = insert_conversation(
-                question=prompt,
-                response=response,
-                citations="",
-                model_name=model_name,
-                response_time=response_time, # seconds
-                correct=None,  # No feedback by default
-                user_id=st.session_state.user_id,
-                common_topics=""
-            )
-
-            # Add conversation to streamlit session state
-            st.session_state.messages.append({
-                "role": "user", 
-                "content": prompt,
-                "conversation_id": conversation_id,
-            })
-            st.session_state.messages.append({
-                "role": "assistant", 
-                "content": response,
-                "conversation_id": conversation_id
-            })
-
-            # # Update user session and rerun streamlit
-            update_user_session(st.session_state.user_id)
-            st.rerun()
+from frontend import streamlit
+from backend.statistics import init_db
 
 # Application entrypoint
 if __name__ == "__main__":
     if os.environ.get("STREAMLIT_RUNNING") == "1":
-        main()
+        init_db()  # Initialize the database 
+        streamlit.main()
     else:
         os.environ["STREAMLIT_RUNNING"] = "1"
         subprocess.run(["streamlit", "run", __file__, "--server.port=5003", "--server.address=0.0.0.0", "--server.baseUrlPath=/team3"])
