+++ data/default/faiss_indexes/pdf_collection_faiss_index/index.faiss
