--- nginx.conf
@@ -1,19 +1,0 @@
-events {}
-
-http {
-    server {
-        listen 83;  # Listen on port 83 for HTTP requests
-
-        location /team3/ {
-            rewrite ^/team3/(.*) /$1 break;  # Rewrite the URL to strip out the subpath for proper routing
-            proxy_pass http://localhost:5003/;  # Proxy requests to the Streamlit app running on port 5003
-            proxy_http_version 1.1;
-            proxy_set_header Upgrade $http_upgrade;  # Allow for websockets
-            proxy_set_header Connection "upgrade";  # Allow for websockets
-            proxy_set_header Host $host;  # Pass the host header
-            proxy_set_header X-Real-IP $remote_addr;  # Pass the real IP of the client
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;  # Forward the original client IP
-            proxy_set_header X-Forwarded-Proto $scheme;  # Forward the protocol used
-        }
-    }
-}
